fix: create adaptor for uint128_t

+++ barretenberg/cpp/src/barretenberg/vm2/common/aztec_types.hpp
@@ -112,8 +112,8 @@ struct PublicDataWrite {
 ////////////////////////////////////////////////////////////////////////////
 
 struct GasFees {
-    uint64_t feePerDaGas;
-    uint64_t feePerL2Gas;
+    uint128_t feePerDaGas;
+    uint128_t feePerL2Gas;
 
     bool operator==(const GasFees& other) const = default;
